+++ karma.config.js
@@ -3,6 +3,9 @@ const path = require('path');
 const srcPath = path.join(__dirname, 'src');
 const testPath = path.join(__dirname, 'test');
 
+// Set environment to testing
+process.env.NODE_ENV = 'testing';
+
 module.exports = function (config) {
     config.set({
         basePath: './',