+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## Changelog
 
 
+* **0.2.3** Upgrading to latest libraries
+    - Major
+        - React 15.0.1
+        - Fabric 1.6.0
+    - Other 
+        - babel
+        - react-color 
+        - etc.           
 * **0.2.2** Minor bug fixes and improvements release
 * **0.2.1** Bundling fabric to the library, to avoid complexity by adding it as external dependency
 * **0.2.0** Adding FabricJS as the core of the sketch field